+++ packages/docusaurus-mdx-loader/src/rightToc/__tests__/index.test.js
@@ -4,6 +4,7 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
+
 import '@babel/polyfill';
 import {join} from 'path';
 import remark from 'remark';