+++ packages/docusaurus-plugin-content-pages/src/index.js
@@ -17,7 +17,7 @@ const DEFAULT_OPTIONS = {
 };
 
 class DocusaurusPluginContentPages {
-  constructor(opts, context) {
+  constructor(context, opts) {
     this.options = {...DEFAULT_OPTIONS, ...opts};
     this.context = context;
     this.contentPath = path.resolve(this.context.siteDir, this.options.path);