+++ packages/docusaurus/lib/client/prefetch.js
@@ -6,12 +6,12 @@
  */
 
 const support = function(feature) {
-  if (typeof document === `undefined`) {
+  if (typeof document === 'undefined') {
     return false;
   }
-  const fakeLink = document.createElement(`link`);
+  const fakeLink = document.createElement('link');
   try {
-    if (fakeLink.relList && typeof fakeLink.relList.supports === `function`) {
+    if (fakeLink.relList && typeof fakeLink.relList.supports === 'function') {
       return fakeLink.relList.supports(feature);
     }
   } catch (err) {
@@ -22,21 +22,21 @@ const support = function(feature) {
 
 const linkPrefetchStrategy = function(url) {
   return new Promise((resolve, reject) => {
-    if (typeof document === `undefined`) {
+    if (typeof document === 'undefined') {
       reject();
       return;
     }
 
-    const link = document.createElement(`link`);
-    link.setAttribute(`rel`, `prefetch`);
-    link.setAttribute(`href`, url);
+    const link = document.createElement('link');
+    link.setAttribute('rel', 'prefetch');
+    link.setAttribute('href', url);
 
     link.onload = resolve;
     link.onerror = reject;
 
     const parentElement =
-      document.getElementsByTagName(`head`)[0] ||
-      document.getElementsByName(`script`)[0].parentNode;
+      document.getElementsByTagName('head')[0] ||
+      document.getElementsByName('script')[0].parentNode;
     parentElement.appendChild(link);
   });
 };
@@ -44,7 +44,7 @@ const linkPrefetchStrategy = function(url) {
 const xhrPrefetchStrategy = function(url) {
   return new Promise((resolve, reject) => {
     const req = new XMLHttpRequest();
-    req.open(`GET`, url, true);
+    req.open('GET', url, true);
     req.withCredentials = true;
 
     req.onload = () => {
@@ -59,7 +59,7 @@ const xhrPrefetchStrategy = function(url) {
   });
 };
 
-const supportedPrefetchStrategy = support(`prefetch`)
+const supportedPrefetchStrategy = support('prefetch')
   ? linkPrefetchStrategy
   : xhrPrefetchStrategy;